backend/requirements.txt
django==4.1.6
psycopg2==2.9.5
model-bakery==1.10.1
coverage==7.2.1
djangorestframework==3.14.0
factory-boy==3.3.0
black==23.11.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==3.14.0
